Make Oracle 11g database ready output compatible with Oracle 21c
Till now, in various places we are using the "Database opened."
output to determine when the Oracle database is up and accepting
connections. We use that, for example, in moodle-docker:

https://github.com/moodlehq/moodle-docker/blob/master/bin/moodle-docker-wait-for-db#L17

But, the init scripts with Oracle 21c are more complex and require
to restart the database once more, so that strings isn't a good
candidate any more, because the first start is detected and the
wait script returns too early.

So, instead, what we are doing here is to ensure that the Oracle 11g
version does add "DATABASE IS READY TO USE!" at the end of the
init, because that exact string is also output by the 21c version.

That way we have a consistent string to be be checked no matter the
version of the Oracle database.