Prepare both the XE (CDB) and XEPDB1 (PDB) to be able to use any
Right now, mainly for compatibility with older Oracle versions,
we are using the "XE" database the comes with Oracle 21c. But that's
the container / root (CDB) database and it doesn't make much sense
to create stuff there. Specially because dictionary / metadata
access is noticeably slower (CDBs contains information for ALL
databases in the system).

Hence, we are also enabling the default pluggable database (PDB)
with name "XEPDB1", to be ready for a Moodle installation. It should
be quicker than the CDB one because of the reasons mentioned above.

And every little improvement that we get with Oracle is
welcome, coz its slowness makes it much, much prone to random
errors, specially with Behat.

Also, adjust a little bit the README information.