Prepare both the XE (CDB) and XEPDB1 (PDB) to be able to use any
Right now, mainly for compatibility with older Oracle versions,
we are using the "XE" database the comes with Oracle 21c. But that's
the container / root (CDB) database and it doesn't make much sense
to create stuff there. Specially because dictionary / metadata
access is noticeably slower (CDBs contains information for ALL
databases in the system).

Hence, we are also enabling the default pluggable database (PDB)
with name "XEPDB1", to be ready for a Moodle installation. It should
be quicker than the CDB one because of the reasons mentioned above.

And every little improvement that we get with Oracle is
welcome, coz its slowness makes it much, much prone to random
errors, specially with Behat.

Also, adjust a little bit the README information.

And, finally, and not less important, let's add to Dockerfile:

RUN chmod 777 -R /opt/oracle/oradata

Because, without that, tmpfs mounts don't work ok (because the
unzio utility gets permission denied to unzip there). If that works
ok, I'll create an issue upstream to see if we can get that done
in the base images.